% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{AddFields}
\alias{AddFields}
\title{Add fields to the MLS Data Data Frame
 @projEnv Project Environment
 @return None
#export}
\usage{
AddFields(projEnv)
}
\description{
Add fields to the MLS Data Data Frame
 @projEnv Project Environment
 @return None
#export
}
