% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage_1.R
\name{allowedInteractions}
\alias{allowedInteractions}
\title{Filter for allowing interactions}
\usage{
allowedInteractions(degree, pred, parents, namesx)
}
\description{
@projEnv    Project Environment
 @return     TRUE/FALSE
}
\details{
@example
 This is called by R/earth

This is the "allowed" function for R:earth that returns TRUE for supported
interactions
}
