% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage_1.R
\name{callEarth}
\alias{callEarth}
\title{Call earth
 @projEnv    Project Environment
 @return     earth output}
\usage{
callEarth(projEnv)
}
\description{
@example
 findContribVars(projEnv,["3107.608","-0.2457188*SaleAge","-2.02277e-05*pmax(0,LatScaled - -444488)*pmax(0, Longitude - -328751)"] )
 returns [["SaleAge"],["LatScaled","Longitude"]] which would indicate the model will provide contributions for SaleAge towards
 SalePrice_K (the taraget variable) and the interaction between LatScaled/Longitude and SalePrice_K
}
