% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{CompressColumnNames}
\alias{CompressColumnNames}
\title{Compress/clean column names
 Replace some characters with more descriptive names and remove unwanted characters such as spaces and periods
 @projEnv Name vector
 @return  Name vector}
\usage{
CompressColumnNames(cnms)
}
\description{
Compress/clean column names
 Replace some characters with more descriptive names and remove unwanted characters such as spaces and periods
 @projEnv Name vector
 @return  Name vector
}
