% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{CompressColumnNamesAllDF}
\alias{CompressColumnNamesAllDF}
\title{Apply CleanColumNames() to all project Data Frames
 @projEnv Name vector
 @return  Name vector}
\usage{
CompressColumnNamesAllDF(projEnv)
}
\description{
Apply CleanColumNames() to all project Data Frames
 @projEnv Name vector
 @return  Name vector
}
