% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KVEsfr_Utilities.R
\name{ConvertBoolenToBinary}
\alias{ConvertBoolenToBinary}
\title{Convert a boolean value to 1 if true, otherwise 0}
\usage{
ConvertBoolenToBinary(boolVal)
}
\description{
Convert a boolean value to 1 if true, otherwise 0
}
