% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{ConvertDatesToChar}
\alias{ConvertDatesToChar}
\title{We need to convert Data Frame Dates to Char/string for the upload to SQLite to work correctly
 @projEnv Name vector
 @return  Name vector}
\usage{
ConvertDatesToChar(projEnv, df)
}
\description{
We need to convert Data Frame Dates to Char/string for the upload to SQLite to work correctly
 @projEnv Name vector
 @return  Name vector
}
