% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{ConvertDatesToCharAllDF}
\alias{ConvertDatesToCharAllDF}
\title{We need to convert Data Frame Dates to Char/string for the upload to SQLite to work correctly
 @projEnv Project Environment
 @return  Project Environment}
\usage{
ConvertDatesToCharAllDF(projEnv)
}
\description{
We need to convert Data Frame Dates to Char/string for the upload to SQLite to work correctly
 @projEnv Project Environment
 @return  Project Environment
}
