% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KVEsfr_Utilities.R
\name{ConvertDfFieldBoolToBinary}
\alias{ConvertDfFieldBoolToBinary}
\title{Comresses column names for MLSData Data Frame and then
converts NA feilds to FALSE and Boolean fields to binary}
\usage{
ConvertDfFieldBoolToBinary(projEnv, fldName)
}
\description{
Comresses column names for MLSData Data Frame and then
converts NA feilds to FALSE and Boolean fields to binary
}
