% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{DoCalculations1}
\alias{DoCalculations1}
\title{Do the calculations for calculated fields}
\usage{
DoCalculations1(projEnv)
}
\value{
None
#export
}
\description{
Do the calculations for calculated fields
}
