% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage_1.R
\name{findContribVars}
\alias{findContribVars}
\title{Find the unique set of contribution variables and interactions.  These may be one-way with only one variable name
 or two or more way to represent interactions.  The variable names will be sorted alphabetically
 and the final list of vectors will have duplicates removed.
 @projEnv    Project Environment
 @listModel  This is a vector of the additive components of the model
             Each component will have one or more variable terms combined with constants and operators where
             variables are combined with a multiplicative operator
 @return     A list of unique vectors, where each vector contains the sorted names of the variables used in one
             or more of the component terms  in listModel
 @example
 findContribVars(projEnv,["3107.608","-0.2457188*SaleAge","-2.02277e-05*pmax(0,LatScaled - -444488)*pmax(0, Longitude - -328751)"] )
 returns [["SaleAge"],["LatScaled","Longitude"]] which would indicate the model will provide contributions for SaleAge towards
 SalePrice_K (the taraget variable) and the interaction between LatScaled/Longitude and SalePrice_K}
\usage{
findContribVars(projEnv, listModel)
}
\description{
Find the unique set of contribution variables and interactions.  These may be one-way with only one variable name
 or two or more way to represent interactions.  The variable names will be sorted alphabetically
 and the final list of vectors will have duplicates removed.
 @projEnv    Project Environment
 @listModel  This is a vector of the additive components of the model
             Each component will have one or more variable terms combined with constants and operators where
             variables are combined with a multiplicative operator
 @return     A list of unique vectors, where each vector contains the sorted names of the variables used in one
             or more of the component terms  in listModel
 @example
 findContribVars(projEnv,["3107.608","-0.2457188*SaleAge","-2.02277e-05*pmax(0,LatScaled - -444488)*pmax(0, Longitude - -328751)"] )
 returns [["SaleAge"],["LatScaled","Longitude"]] which would indicate the model will provide contributions for SaleAge towards
 SalePrice_K (the taraget variable) and the interaction between LatScaled/Longitude and SalePrice_K
}
