% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{LoadMappingsExcelToDataFrame}
\alias{LoadMappingsExcelToDataFrame}
\title{Load the Mappings (or configuation file) to a Data Frame
 @projEnv Project Environment
 @return  Project Environment}
\usage{
LoadMappingsExcelToDataFrame(projEnv)
}
\description{
Load the Mappings (or configuation file) to a Data Frame
 @projEnv Project Environment
 @return  Project Environment
}
