% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{LoadMlsExcelToDataFrame}
\alias{LoadMlsExcelToDataFrame}
\title{Loads the initial MLS Excel Data File to Data Frames
 @projEnv Project Environment
 @return  Project Environment}
\usage{
LoadMlsExcelToDataFrame(projEnv)
}
\description{
#export
}
