% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage_1.R
\name{p1_initialize}
\alias{p1_initialize}
\title{Filter for allowing interactions
 @projEnv    Project Environment
 @return     TRUE/FALSE}
\usage{
p1_initialize(projEnv)
}
\description{
@example
 p1_initiallize
}
\details{
p1_initialize loads the Excel Interaction Matrix and uses it to create
a table for deciding which interactions to allow
}
