% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{RenameFieldsAndSetAvailable}
\alias{RenameFieldsAndSetAvailable}
\title{Do any field renaming as set in the Mappings:MlsMapping sheet and then print as the initial AvailableFields in same named sheet}
\usage{
RenameFieldsAndSetAvailable(projEnv)
}
\value{
None
}
\description{
Do any field renaming as set in the Mappings:MlsMapping sheet and then print as the initial AvailableFields in same named sheet
}
