% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{RepPeriodWithSpaceColumnNames}
\alias{RepPeriodWithSpaceColumnNames}
\title{Replace periods in name vector with spaces
 @projEnv Name vector
 @return  Name vector}
\usage{
RepPeriodWithSpaceColumnNames(cnms)
}
\description{
Replace periods in name vector with spaces
 @projEnv Name vector
 @return  Name vector
}
