% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{SetUp_0}
\alias{SetUp_0}
\title{SetUp
 @projEnv Project Environment
 @codeFolder codeFolder,
 @projectParentFolder Project files parent folder
 @projectID Project ID
 @mlsData MLS Data File
 @mlsSheet MLS Data File Sheet Name
 @subjectCQA  Subject CQA}
\usage{
SetUp_0(
  projEnv,
  codeFolder,
  projectParentFolder,
  projectID,
  mlsData,
  mlsSheet,
  targetVariable,
  bootstrap,
  pmethod,
  ncross,
  nfold,
  keepxy,
  penalty_1D,
  penalty_2D,
  trace,
  maxterms,
  nprune,
  minspan,
  endspan,
  degree,
  subjectCQA
)
}
\description{
@example
 Stage_1(projEnv)
}
\details{
Out main Stage I function:
#export
}
