% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{SetUpProjectEnvironment}
\alias{SetUpProjectEnvironment}
\title{Initial call is here to set up the environment under projEnv
 @projEnv Project Environment
 @codeFolder codeFolder,
 @projectParentFolder Project files parent folder
 @projectID Project ID
 @mlsDataFile MLS Data File
 @mlsSheet
 @targetVariable
 @pmethod
 @maxterms
 @mlsSheet MLS Data File Sheet Name
 @minspan
 @endspan
 @degree Earth degree arg
 @subjectCQA  Subject CQA
Initial call is here to set up the environment under projEnv
#export}
\usage{
SetUpProjectEnvironment(
  projEnv,
  codeFolder,
  projectParentFolder,
  projectID,
  mlsDataFile,
  mlsSheet,
  targetVariable,
  bootstrap,
  pmethod,
  ncross,
  nfold,
  keepxy,
  penalty_1D,
  penalty_2D,
  trace,
  maxterms,
  nprune,
  minspan,
  endspan,
  degree,
  subjectCQA
)
}
\description{
Initial call is here to set up the environment under projEnv
 @projEnv Project Environment
 @codeFolder codeFolder,
 @projectParentFolder Project files parent folder
 @projectID Project ID
 @mlsDataFile MLS Data File
 @mlsSheet
 @targetVariable
 @pmethod
 @maxterms
 @mlsSheet MLS Data File Sheet Name
 @minspan
 @endspan
 @degree Earth degree arg
 @subjectCQA  Subject CQA
Initial call is here to set up the environment under projEnv
#export
}
