% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage_1.R
\name{Stage_1}
\alias{Stage_1}
\title{Stage I}
\usage{
Stage_1(projEnv)
}
\description{
@projEnv    Project Environment
 @return     None
}
\details{
@example
 Stage_1(projEnv)

Out main Stage I function:
#export
}
