% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stage_2.R
\name{Stage_2}
\alias{Stage_2}
\title{Stage2}
\usage{
Stage_2(projEnv)
}
\description{
@projEnv    Project Environment
 @return     None
}
\details{
@example
 Stage_2(projEnv)

Out main Stage II function:
#export
}
