% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{WriteConfigMappingsDfToSqlite}
\alias{WriteConfigMappingsDfToSqlite}
\title{Write Config Mappings To SQLite
 @projEnv Name vector
 @return  Name vector}
\usage{
WriteConfigMappingsDfToSqlite(projEnv)
}
\description{
Write Config Mappings To SQLite
 @projEnv Name vector
 @return  Name vector
}
