% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUp_0.R
\name{WriteMlsDfToSqlite}
\alias{WriteMlsDfToSqlite}
\title{Store the MLS Data Frame to SQLite for ease of use, reliability}
\usage{
WriteMlsDfToSqlite(projEnv)
}
\description{
Store the MLS Data Frame to SQLite for ease of use, reliability
}
